Fix syntax error in autograder yaml
You had a yaml syntax error in the autograder
file you attempted to create.  Basically yaml doesn't allow the last
item in a list of items to end with a `,`, so the last comma had to
be removed.  Fixed this to see if tests you made are passing, though
I checked out the code and ran by hand so I know the tests you wrote
are passing.